Removed double types

+++ Source/DigitalRise.Graphics/Misc/Ephemeris/Ephemeris_Planets.cs
@@ -41,20 +41,20 @@ namespace DigitalRise.Graphics
     // Stores the parameters required to uniquely identify a specific planet orbit.
     private struct OrbitalElements
     {
-      public double Period;
-      public double EpochLongitude;
-      public double PerihelionLongitude;
-      public double Eccentricity;
-      public double SemiMajorAxis;
-      public double Inclination;
-      public double LongitudeAscendingNode;
-      public double AngularDiameter;
-      public double VisualMagnitude;
+      public float Period;
+      public float EpochLongitude;
+      public float PerihelionLongitude;
+      public float Eccentricity;
+      public float SemiMajorAxis;
+      public float Inclination;
+      public float LongitudeAscendingNode;
+      public float AngularDiameter;
+      public float VisualMagnitude;
 
 
-      public OrbitalElements(double period, double epochLongitude, double perihelionLongitude,
-        double eccentricity, double semiMajorAxis, double inclination, double longitudeAscendingNode,
-        double angularDiameter, double visualMagnitude)
+      public OrbitalElements(float period, float epochLongitude, float perihelionLongitude,
+        float eccentricity, float semiMajorAxis, float inclination, float longitudeAscendingNode,
+        float angularDiameter, float visualMagnitude)
       {
         Period = period;
         EpochLongitude = epochLongitude;
@@ -73,12 +73,12 @@ namespace DigitalRise.Graphics
     //private struct PlanetData
     //{
     //  // Equatorial coordinates.
-    //  public double RightAscension;
-    //  public double Declination;
+    //  public float RightAscension;
+    //  public float Declination;
 
     //  // Measures the brightness of the planet as seen on earth. Lower values mean more brightness.
     //  // (I think this is the same as "apparent magnitude".)
-    //  public double VisualMagnitude;
+    //  public float VisualMagnitude;
     //};
     #endregion
 
@@ -93,28 +93,28 @@ namespace DigitalRise.Graphics
     private readonly OrbitalElements[] _planetElements =
     {
       // Mercury
-      new OrbitalElements(0.240852, MathHelper.ToRadians(60.750646), MathHelper.ToRadians(77.299833), 
-        0.205633, 0.387099, MathHelper.ToRadians(7.004540), MathHelper.ToRadians(48.212740), 6.74, -0.42),
+      new OrbitalElements(0.240852f, MathHelper.ToRadians(60.750646f), MathHelper.ToRadians(77.299833f), 
+        0.205633f, 0.387099f, MathHelper.ToRadians(7.004540f), MathHelper.ToRadians(48.212740f), 6.74f, -0.42f),
 
       // Venus
-      new OrbitalElements(0.615211, MathHelper.ToRadians(88.455855), MathHelper.ToRadians(131.430236), 
-        0.006778, 0.723332, MathHelper.ToRadians(3.394535), MathHelper.ToRadians(76.589820), 16.92, -4.40),
+      new OrbitalElements(0.615211f, MathHelper.ToRadians(88.455855f), MathHelper.ToRadians(131.430236f), 
+        0.006778f, 0.723332f, MathHelper.ToRadians(3.394535f), MathHelper.ToRadians(76.589820f), 16.92f, -4.40f),
 
       // Earth
-      new OrbitalElements(1.00004, MathHelper.ToRadians(99.403308), MathHelper.ToRadians(102.768413), 
-        0.016713, 1.00000, 0, 0, 0, 0),
+      new OrbitalElements(1.00004f, MathHelper.ToRadians(99.403308f), MathHelper.ToRadians(102.768413f), 
+        0.016713f, 1.00000f, 0, 0, 0, 0),
 
       // Mars
-      new OrbitalElements(1.880932, MathHelper.ToRadians(240.739474), MathHelper.ToRadians(335.874939), 
-        0.093396, 1.523688, MathHelper.ToRadians(1.849736), MathHelper.ToRadians(49.480308), 9.36, -1.52),
+      new OrbitalElements(1.880932f, MathHelper.ToRadians(240.739474f), MathHelper.ToRadians(335.874939f), 
+        0.093396f, 1.523688f, MathHelper.ToRadians(1.849736f), MathHelper.ToRadians(49.480308f), 9.36f, -1.52f),
 
       // Jupiter
-      new OrbitalElements(11.863075, MathHelper.ToRadians(90.638185), MathHelper.ToRadians(14.170747), 
-        0.048482, 5.202561, MathHelper.ToRadians(1.303613), MathHelper.ToRadians(100.353142), 196.74, -9.40),
+      new OrbitalElements(11.863075f, MathHelper.ToRadians(90.638185f), MathHelper.ToRadians(14.170747f), 
+        0.048482f, 5.202561f, MathHelper.ToRadians(1.303613f), MathHelper.ToRadians(100.353142f), 196.74f, -9.40f),
 
       // Saturn
-      new OrbitalElements(29.471362, MathHelper.ToRadians(287.690033), MathHelper.ToRadians(92.861407), 
-        0.055581, 9.554747, MathHelper.ToRadians(2.488980), MathHelper.ToRadians(113.576139), 165.60, -8.88)
+      new OrbitalElements(29.471362f, MathHelper.ToRadians(287.690033f), MathHelper.ToRadians(92.861407f), 
+        0.055581f, 9.554747f, MathHelper.ToRadians(2.488980f), MathHelper.ToRadians(113.576139f), 165.60f, -8.88f)
     };
     #endregion
 
@@ -148,34 +148,34 @@ namespace DigitalRise.Graphics
     //{
     //  int index = (int)planet;
 
-    //  double Np = ((2.0 * ConstantsD.Pi) / 365.242191) * (_epoch1990Days / planetElements[index].Period);
+    //  float Np = ((2.0 * ConstantsF.Pi) / 365.242191) * (_epoch1990Days / planetElements[index].Period);
     //  Np = InRange(Np);
 
-    //  double Mp = Np + planetElements[index].EpochLongitude - planetElements[index].PerihelionLongitude;
+    //  float Mp = Np + planetElements[index].EpochLongitude - planetElements[index].PerihelionLongitude;
 
-    //  double heliocentricLongitude = Np + 2.0 * planetElements[index].Eccentricity * Math.Sin(Mp) +
+    //  float heliocentricLongitude = Np + 2.0 * planetElements[index].Eccentricity * Math.Sin(Mp) +
     //                                 planetElements[index].EpochLongitude;
     //  heliocentricLongitude = InRange(heliocentricLongitude);
 
-    //  double vp = heliocentricLongitude - planetElements[index].PerihelionLongitude;
+    //  float vp = heliocentricLongitude - planetElements[index].PerihelionLongitude;
 
-    //  double radius = (planetElements[index].SemiMajorAxis
+    //  float radius = (planetElements[index].SemiMajorAxis
     //              * (1.0 - planetElements[index].Eccentricity * planetElements[index].Eccentricity))
     //           / (1.0 + planetElements[index].Eccentricity * Math.Cos(vp));
 
-    //  double heliocentricLatitude = Math.Asin(Math.Sin(heliocentricLongitude -
+    //  float heliocentricLatitude = Math.Asin(Math.Sin(heliocentricLongitude -
     //                                         planetElements[index].LongitudeAscendingNode) * Math.Sin(planetElements[index].Inclination));
     //  heliocentricLatitude = InRange(heliocentricLatitude);
 
-    //  double y = Math.Sin(heliocentricLongitude - planetElements[index].LongitudeAscendingNode) *
+    //  float y = Math.Sin(heliocentricLongitude - planetElements[index].LongitudeAscendingNode) *
     //             Math.Cos(planetElements[index].Inclination);
-    //  double x = Math.Cos(heliocentricLongitude - planetElements[index].LongitudeAscendingNode);
+    //  float x = Math.Cos(heliocentricLongitude - planetElements[index].LongitudeAscendingNode);
 
-    //  double projectedHeliocentricLongitude = Math.Atan2(y, x) + planetElements[index].LongitudeAscendingNode;
+    //  float projectedHeliocentricLongitude = Math.Atan2(y, x) + planetElements[index].LongitudeAscendingNode;
 
-    //  double projectedRadius = radius * Math.Cos(heliocentricLatitude);
+    //  float projectedRadius = radius * Math.Cos(heliocentricLatitude);
 
-    //  double eclipticLongitude;
+    //  float eclipticLongitude;
 
     //  if (index > (int)VisiblePlanets.Earth)
     //  {
@@ -185,29 +185,29 @@ namespace DigitalRise.Graphics
     //  }
     //  else
     //  {
-    //    eclipticLongitude = ConstantsD.Pi + _earthEclipticLongitude + Math.Atan((projectedRadius * Math.Sin(_earthEclipticLongitude - projectedHeliocentricLongitude))
+    //    eclipticLongitude = ConstantsF.Pi + _earthEclipticLongitude + Math.Atan((projectedRadius * Math.Sin(_earthEclipticLongitude - projectedHeliocentricLongitude))
     //                                       / (_earthRadius - projectedRadius * Math.Cos(_earthEclipticLongitude - projectedHeliocentricLongitude)));
     //  }
 
     //  eclipticLongitude = InRange(eclipticLongitude);
 
-    //  double eclipticLatitude = Math.Atan((projectedRadius * Math.Tan(heliocentricLatitude)
+    //  float eclipticLatitude = Math.Atan((projectedRadius * Math.Tan(heliocentricLatitude)
     //                                   * Math.Sin(eclipticLongitude - projectedHeliocentricLongitude))
     //                                  / (_earthRadius * Math.Sin(projectedHeliocentricLongitude - _earthEclipticLongitude)));
 
-    //  double ra = Math.Atan2((Math.Sin(eclipticLongitude) * Math.Cos(_e) - Math.Tan(eclipticLatitude) * Math.Sin(_e))
+    //  float ra = Math.Atan2((Math.Sin(eclipticLongitude) * Math.Cos(_e) - Math.Tan(eclipticLatitude) * Math.Sin(_e))
     //                     , Math.Cos(eclipticLongitude));
 
-    //  double dec = Math.Asin(Math.Sin(eclipticLatitude) * Math.Cos(_e) + Math.Cos(eclipticLatitude) * Math.Sin(_e)
+    //  float dec = Math.Asin(Math.Sin(eclipticLatitude) * Math.Cos(_e) + Math.Cos(eclipticLatitude) * Math.Sin(_e)
     //                    * Math.Sin(eclipticLongitude));
 
-    //  double dist2 = _earthRadius * _earthRadius + radius * radius - 2 * _earthRadius * radius * Math.Cos(heliocentricLongitude - _earthEclipticLongitude);
-    //  double dist = Math.Sqrt(dist2);
+    //  float dist2 = _earthRadius * _earthRadius + radius * radius - 2 * _earthRadius * radius * Math.Cos(heliocentricLongitude - _earthEclipticLongitude);
+    //  float dist = Math.Sqrt(dist2);
 
-    //  double d = eclipticLongitude - heliocentricLongitude;
-    //  double phase = 0.5 * (1.0 + Math.Cos(d));
+    //  float d = eclipticLongitude - heliocentricLongitude;
+    //  float phase = 0.5 * (1.0 + Math.Cos(d));
 
-    //  double visualMagnitude;
+    //  float visualMagnitude;
 
     //  if (index == (int)VisiblePlanets.Venus)
     //  {
@@ -232,7 +232,7 @@ namespace DigitalRise.Graphics
     ///// <param name="rightAscension">The right ascension of .</param>
     ///// <param name="declination">The declination.</param>
     ///// <param name="visualMagnitude">The visual magnitude (lower values are brighter).</param>
-    //public void GetPlanetPosition(VisiblePlanets planet, out double rightAscension, out double declination, out double visualMagnitude)
+    //public void GetPlanetPosition(VisiblePlanets planet, out float rightAscension, out float declination, out float visualMagnitude)
     //{
     //  if (planet == VisiblePlanets.Earth)
     //    throw new ArgumentException("This method must not be called for the earth.");
